refactor(ui): Refactor product overview page
In order to prevent full re-render of the product overview table, when
the repositories table is navigated, move the table to a separate
table wrapper component, which handles its own data loading, table
rendering and handling of its loading state.

Resolves #1606.
